feat: split dashboard into separate route with cached landing
- Add proxy.ts for Next.js 16 auth-based routing (/ → /dashboard)
- Simplify app/page.tsx to cached static landing page
- Create app/dashboard/page.tsx with Suspense boundary for dynamic content
- Fix blocking route error by isolating dynamic session checks
